// src/features/Controls/ChordDetailsControl/enums.ts
export enum ChordVisibility {
  HIDDEN = 'hidden',
  UNSET = 'unset',
}

export enum HSystem {
  WORLD = 'world',
  CZECH = 'czech',
}

export enum Notes {
  C = 'C',
  C_SHARP = 'C#',
  D = 'D',
  D_SHARP = 'D#',
  E = 'E',
  F = 'F',
  F_SHARP = 'F#',
  G = 'G',
  G_SHARP = 'G#',
  A = 'A',
  A_SHARP = 'A#',
  B = 'B',
}